Additional Feedback and Requirements from MYOB
Added more use of Java 8 constructs. Made the rest service more restful
by implementing HATEOS. Separated out the payment period logic so the
single responsibily principle was adhered to. Removed a lot of the
logging and code commenting that was in place purely for the assignment,
as well as the assumptions and approaches, let the code speak for
itself.